requirements.txt
Flask==2.3.2
gunicorn==21.2.0
Jinja2==3.1.2
requests==2.32.3
python-dotenv==1.0.1

